Disable toggle for forcing apps to be resizable in user builds
This can potentially cause problems for app developers that have this
manifest value disabled. If an app is not meant to be resized, is for
a good reason. The main reason is usually the app's layout does not
function correctly in multi-window. The user should not be allowed to
toggle this on while using a user build.

If the user or another developer insists on having this option, they can
run a userdebug build and have at it.

Original commit below and credit given to ngc4622 to avoid issues
https://github.com/LineageOS/android_packages_apps_Settings/commit/60cd0481c429f7f4e33f5d2aa0ec18e3a70afc5c

Change-Id: Id8331d2f923091e6fe30a81ba3592085c489cfcc